requirements.txt
pillow
streamlit
python-magic